feat: consolidate the billing page across tiers
Pay as you go now shares the enterprise tokens-under-management usage view
(usage card, breakdown chart, details table) with its Stripe invoice estimate
at the head of the section. Sections are ordered Usage, Inference caps, Plan,
Billing notifications, and organizations with no recorded usage get an
explicit empty state instead of a blank chart. The duplicate inference cap
meters in the old PAYG usage section are gone — the caps section directly
below already reports the same spend.
