+++ client/dashboard/src/components/billing/inference-cap-meter.tsx
@@ -1,7 +1,5 @@
 import {
   crossedSpendCapThreshold,
-  inferenceCapBillingNote,
-  inferenceCapInvoiceNote,
   inferenceCapLabel,
   type SpendCapThreshold,
   spendCapFillPercent,
@@ -25,49 +23,26 @@ const usdMeter = new Intl.NumberFormat("en-US", {
  * This calendar month's spend against one Gram-managed inference cap.
  *
  * Shared by every surface that reports a cap, so none of them can disagree
- * about what has been spent, or about which band the spend is in.
- *
- * `title` is what a caller that names the cap itself doesn't want: the cap's
- * own control already names it in the label of the field that sets it, and
- * repeating it directly below reads as two separate things. `billingNote` is
- * the same split: a meter that sits under the invoice estimate has to say how
- * these figures relate to it, while the control is nowhere near one.
- *
- * The note is picked here rather than passed in because which half of it is
- * true depends on the branch below — copy that talks about the cap's month is
- * a contradiction on a key that has no cap.
+ * about what has been spent, or about which band the spend is in. It never
+ * names the cap: every caller sits under the control or label that already
+ * does, and repeating the name directly below reads as two separate things.
  */
 export function InferenceCapMeter({
   cap,
-  title = true,
-  billingNote = false,
 }: {
   cap: InferenceSpendCap;
-  title?: boolean;
-  billingNote?: boolean;
 }): JSX.Element {
   const label = inferenceCapLabel(cap.keyType);
   const spent = usdMeter.format(cap.creditsUsed);
-  const heading = title ? <Text className="font-medium">{label}</Text> : null;
 
   // Without a cap the spend has nothing to be a proportion of, and a full-width
   // bar would read as a limit that was reached. The figure still shows: it is
-  // the only place this month's spend on this key appears. So does the invoice
-  // half of the billing note — an uncapped key still spends money, and whether
-  // that money reaches the invoice is exactly what the note is there to say.
-  // Only that half: the rest of it is about the cap's month rolling over, which
-  // would contradict the "No cap is set." printed directly above.
+  // the only place this month's spend on this key appears.
   if (!(cap.monthlyCredits > 0)) {
     return (
-      <Stack gap={1}>
-        {heading}
-        <Text muted small className="tabular-nums">
-          {spent} spent this month. No cap is set.
-        </Text>
-        <Footnote
-          text={billingNote ? inferenceCapInvoiceNote(cap.keyType) : ""}
-        />
-      </Stack>
+      <Text muted small className="tabular-nums">
+        {spent} spent this month. No cap is set.
+      </Text>
     );
   }
 
@@ -77,21 +52,13 @@ export function InferenceCapMeter({
   );
   const percent = spendCapFillPercent(cap.creditsUsed, cap.monthlyCredits);
   const limit = usdMeter.format(cap.monthlyCredits);
-  const footnote = [
-    capNote(threshold, cap.disabled),
-    billingNote ? inferenceCapBillingNote(cap.keyType) : "",
-  ]
-    .filter(Boolean)
-    .join(" ");
+  const footnote = capNote(threshold, cap.disabled);
 
   return (
     <Stack gap={2}>
-      <Stack direction="horizontal" justify="space-between" gap={3}>
-        {heading}
-        <Text muted small className="tabular-nums">
-          {spent} of {limit}
-        </Text>
-      </Stack>
+      <Text muted small className="tabular-nums">
+        {spent} of {limit}
+      </Text>
       <div
         role="progressbar"
         aria-label={`${label}: ${spent} of the ${limit} monthly cap`}
@@ -108,24 +75,14 @@ export function InferenceCapMeter({
           style={{ width: `${percent}%` }}
         />
       </div>
-      <Footnote text={footnote} />
+      {/* Below the first threshold there is no note, and an empty line must
+          not take up space. */}
+      {footnote !== "" && (
+        <Text muted small>
+          {footnote}
+        </Text>
+      )}
     </Stack>
-  );
-}
-
-/**
- * The line under a meter, dropped entirely when there is nothing to say.
- *
- * Both branches render one, and both can end up with an empty string — below
- * the first threshold there is no threshold note, and the cap's own control
- * asks for no billing note at all — so an empty line must not take up space.
- */
-function Footnote({ text }: { text: string }): JSX.Element | null {
-  if (text === "") return null;
-  return (
-    <Text muted small>
-      {text}
-    </Text>
   );
 }
